# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chimikee API",
    "description": "Chimikee API",
    "contact": {
      "name": "Softup LLC",
      "email": "info@softup.io"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/profile": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fullName": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "gender": {
                  "example": "any"
                },
                "dob": {
                  "example": "any"
                },
                "bio": {
                  "example": "any"
                },
                "username": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/change-password": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "newPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}